+++ EasyPost/Models/API/CarrierFields.cs
@@ -15,7 +15,7 @@ namespace EasyPost.Models.API
         ///     The credentials used in the production environment.
         /// </summary>
         [JsonProperty("credentials")]
-        public Dictionary<string, CredentialsField>? Credentials { get; set; }
+        public Dictionary<string, CarrierField>? Credentials { get; set; }
 
         /// <summary>
         ///     The credentials used in the test environment.
@@ -24,7 +24,7 @@ namespace EasyPost.Models.API
 #pragma warning disable SA1515
         // ReSharper disable once InconsistentNaming
 #pragma warning restore SA1515
-        public Dictionary<string, CredentialsField>? TestCredentials { get; set; }
+        public Dictionary<string, CarrierField>? TestCredentials { get; set; }
 
         /// <summary>
         ///     For USPS, this designates that no credentials are required.
@@ -44,7 +44,7 @@ namespace EasyPost.Models.API
     /// <summary>
     ///     Class representing a single EasyPost <see cref="CarrierAccount"/>'s credentials entry details.
     /// </summary>
-    public class CredentialsField : EasyPostObject
+    public class CarrierField : EasyPostObject
     {
         #region JSON Properties